Refactored `SaloonHelper` to `EventHelper`
